componentização de todos o layout
