+++ src/ui/sections/EducationSection.jsx
@@ -1,0 +1,69 @@
+function EducationSection({t, section, field, renderFields}) {
+    return (
+        <>
+            <div
+                id="formacao"
+                className={`space-y-4 sm:space-y-6 ${section !== "formacao" && "hidden"
+                    }`}
+            >
+                <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center gap-3">
+                    <h2 className="text-xl sm:text-2xl font-semibold text-gray-800 flex items-center">
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-5 w-5 sm:h-6 sm:w-6 mr-2 text-blue-500"
+                            fill="none"
+                            viewBox="0 0 24 24"
+                            stroke="currentColor"
+                        >
+                            <path d="M12 14l9-5-9-5-9 5 9 5z" />
+                            <path d="M12 14l6.16-3.422a12.083 12.083 0 01.665 6.479A11.952 11.952 0 0012 20.055a11.952 11.952 0 00-6.824-2.998 12.078 12.078 0 01.665-6.479L12 14z" />
+                            <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                strokeWidth={2}
+                                d="M12 14l9-5-9-5-9 5 9 5zm0 0l6.16-3.422a12.083 12.083 0 01.665 6.479A11.952 11.952 0 0012 20.055a11.952 11.952 0 00-6.824-2.998 12.078 12.078 0 01.665-6.479L12 14zm-4 6v-7.5l4-2.222"
+                            />
+                        </svg>
+                        {t.campos.formacao}
+                    </h2>
+                    <button
+                        type="button"
+                        onClick={() =>
+                            field("formacoes", {
+                                tipo: "superior",
+                                curso: "",
+                                instituicao: "",
+                                mesInicio: "",
+                                anoInicio: "",
+                                mesFim: "",
+                                anoFim: "",
+                                emAndamento: false,
+                                descricao: "", // Novo campo
+                            })
+                        }
+                        className="flex items-center bg-blue-600 hover:bg-blue-700 text-white text-xs sm:text-sm font-medium px-3 sm:px-4 py-2 rounded-lg transition-colors w-full sm:w-auto justify-center"
+                    >
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-4 w-4 sm:h-5 sm:w-5 mr-1"
+                            viewBox="0 0 20 20"
+                            fill="currentColor"
+                        >
+                            <path
+                                fillRule="evenodd"
+                                d="M10 18a8 8 0 100-16 8 8 0 000 16zm1-11a1 1 0 10-2 0v2H7a1 1 0 100 2h2v2a1 1 0 102 0v-2h2a1 1 0 100-2h-2V7z"
+                                clipRule="evenodd"
+                            />
+                        </svg>
+                        {t.botoes.adicionarFormacao}
+                    </button>
+                </div>
+
+                {renderFields()}
+            </div>
+
+        </>
+    )
+}
+
+export default EducationSection;
