+++ src/ui/sections/ExperiencesSection.jsx
@@ -1,0 +1,85 @@
+function ExperiencesSection({t, section, field, data, renderFields}) {
+    return (
+        <>
+             <div
+                id="experiencia"
+                className={`space-y-4 sm:space-y-6 ${section !== "experiencia" && "hidden"
+                    }`}
+            >
+                <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center gap-3">
+                    <h2 className="text-xl sm:text-2xl font-semibold text-gray-800 flex items-center">
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-5 w-5 sm:h-6 sm:w-6 mr-2 text-blue-500"
+                            fill="none"
+                            viewBox="0 0 24 24"
+                            stroke="currentColor"
+                        >
+                            <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                strokeWidth={2}
+                                d="M21 13.255A23.931 23.931 0 0112 15c-3.183 0-6.22-.62-9-1.745M16 6V4a2 2 0 00-2-2h-4a2 2 0 00-2 2v2m4 6h.01M5 20h14a2 2 0 002-2V8a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"
+                            />
+                        </svg>
+                        {t.campos.experiencia}
+                    </h2>
+                    <button
+                        type="button"
+                        onClick={() =>
+                            field("experiencias", {
+                                cargo: "",
+                                empresa: "",
+                                periodo: "",
+                                tecnologias: "",
+                                atividades: "",
+                                resultados: "",
+                            })
+                        }
+                        className="flex items-center bg-blue-600 hover:bg-blue-700 text-white text-xs sm:text-sm font-medium px-3 sm:px-4 py-2 rounded-lg transition-colors w-full sm:w-auto justify-center"
+                    >
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-4 w-4 sm:h-5 sm:w-5 mr-1"
+                            viewBox="0 0 20 20"
+                            fill="currentColor"
+                        >
+                            <path
+                                fillRule="evenodd"
+                                d="M10 18a8 8 0 100-16 8 8 0 000 16zm1-11a1 1 0 10-2 0v2H7a1 1 0 100 2h2v2a1 1 0 102 0v-2h2a1 1 0 100-2h-2V7z"
+                                clipRule="evenodd"
+                            />
+                        </svg>
+                        {t.botoes.adicionarExperiencia}
+                    </button>
+                </div>
+
+                {data.experiencias.length > 0 ? (
+                    renderFields()
+                ) : (
+                    <div className="bg-gray-50 p-4 sm:p-6 rounded-lg border border-gray-200 text-center">
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-10 w-10 sm:h-12 sm:w-12 mx-auto text-gray-400"
+                            fill="none"
+                            viewBox="0 0 24 24"
+                            stroke="currentColor"
+                        >
+                            <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                strokeWidth={1}
+                                d="M9 13h6m-3-3v6m-9 1V7a2 2 0 012-2h6l2 2h6a2 2 0 012 2v8a2 2 0 01-2 2H5a2 2 0 01-2-2z"
+                            />
+                        </svg>
+                        <p className="mt-2 text-sm text-gray-500">
+                            {t.mensagens.nenhumaExperiencia}
+                        </p>
+                    </div>
+                )}
+            </div>
+        </>
+    )
+}
+
+export default ExperiencesSection;
