+++ src/ui/sections/LanguagesSection.jsx
@@ -1,0 +1,76 @@
+function LanguagesSection({section, t, field, data, renderFields}) {
+    return (
+        <>
+            <div
+                id="idiomas"
+                className={`space-y-4 sm:space-y-6 ${section !== "idiomas" && "hidden"
+                    }`}
+            >
+                <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center gap-3">
+                    <h2 className="text-xl sm:text-2xl font-semibold text-gray-800 flex items-center">
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-5 w-5 sm:h-6 sm:w-6 mr-2 text-blue-500"
+                            fill="none"
+                            viewBox="0 0 24 24"
+                            stroke="currentColor"
+                        >
+                            <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                strokeWidth={2}
+                                d="M3 5h12M9 3v2m1.048 9.5A18.022 18.022 0 016.412 9m6.088 9h7M11 21l5-10 5 10M12.751 5C11.783 10.77 8.07 15.61 3 18.129"
+                            />
+                        </svg>
+                        {t.campos.idiomas}
+                    </h2>
+                    <button
+                        type="button"
+                        onClick={() => field("idiomas", { idioma: "", nivel: "" })}
+                        className="flex items-center bg-blue-600 hover:bg-blue-700 text-white text-xs sm:text-sm font-medium px-3 sm:px-4 py-2 rounded-lg transition-colors w-full sm:w-auto justify-center"
+                    >
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-4 w-4 sm:h-5 sm:w-5 mr-1"
+                            viewBox="0 0 20 20"
+                            fill="currentColor"
+                        >
+                            <path
+                                fillRule="evenodd"
+                                d="M10 18a8 8 0 100-16 8 8 0 000 16zm1-11a1 1 0 10-2 0v2H7a1 1 0 100 2h2v2a1 1 0 102 0v-2h2a1 1 0 100-2h-2V7z"
+                                clipRule="evenodd"
+                            />
+                        </svg>
+                        {t.botoes.adicionarIdioma}
+                    </button>
+                </div>
+
+                {data.idiomas.length > 0 ? (
+                    renderFields()
+                ) : (
+                    <div className="bg-gray-50 p-4 sm:p-6 rounded-lg border border-gray-200 text-center">
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-10 w-10 sm:h-12 sm:w-12 mx-auto text-gray-400"
+                            fill="none"
+                            viewBox="0 0 24 24"
+                            stroke="currentColor"
+                        >
+                            <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                strokeWidth={1}
+                                d="M3 5h12M9 3v2m1.048 9.5A18.022 18.022 0 016.412 9m6.088 9h7M11 21l5-10 5 10M12.751 5C11.783 10.77 8.07 15.61 3 18.129"
+                            />
+                        </svg>
+                        <p className="mt-2 text-sm text-gray-500">
+                            {t.mensagens.nenhumIdioma}
+                        </p>
+                    </div>
+                )}
+            </div>
+        </>
+    )
+}
+
+export default LanguagesSection;
