+++ src/utils/fields.js
@@ -1,389 +1,458 @@
+// campos do curriculo
+export const curriculumFields = {
+  nome: "",
+  telefone: "",
+  ddd: "",
+  codigoPais: "+55",
+  cidade: "",
+  email: "",
+  links: [{ tipo: "linkedin", url: "" }],
+  cargoDesejado: "",
+  resumo: "",
+  experiencias: [],
+  formacoes: [
+    {
+      tipo: "superior",
+      curso: "",
+      instituicao: "",
+      mesInicio: "",
+      anoInicio: "",
+      mesFim: "",
+      anoFim: "",
+      emAndamento: false,
+      descricao: "",
+    },
+  ],
+  habilidades: [],
+  certificacoes: [
+    {
+      titulo: "",
+      emissor: "",
+      data: "",
+      cargaHoraria: "",
+      descricao: "",
+      linkValidacao: "",
+    },
+  ],
+  idiomas: [{ idioma: "", nivel: "" }],
+}
+
 // Opções de idioma para a aplicação
 export const idiomasApp = [
-    { codigo: "pt", nome: "Português", icone: "🇧🇷" },
-    { codigo: "en", nome: "English", icone: "🇺🇸" },
-    { codigo: "es", nome: "Español", icone: "🇪🇸" },
-  ];
+  { codigo: "pt", nome: "Português", icone: "🇧🇷" },
+  { codigo: "en", nome: "English", icone: "🇺🇸" },
+  { codigo: "es", nome: "Español", icone: "🇪🇸" },
+];
 
-  // Textos traduzidos
+// Textos traduzidos
 export const textos = {
-    pt: {
-      tituloApp: "Gerador de Currículo",
-      subtituloApp: "Crie um currículo profissional em minutos",
-      dicasATS: "Dicas para um currículo ATS-friendly",
-      dicasLista: [
-        "Use palavras-chave relevantes para a vaga",
-        "Mantenha o formato simples e legível",
-        "Inclua métricas e resultados concretos",
-        "Destaque suas principais habilidades técnicas",
-      ],
-      campos: {
-        nome: "Informações Pessoais",
-        cargoDesejado: "Cargo Desejado",
-        codigoPais: "Código do País",
-        ddd: "DDD",
-        telefone: "Telefone",
-        cidade: "Cidade",
-        linkedin: "LinkedIn",
-        portfolio: "Portfolio/GitHub",
-        email: "Email*",
-        resumo: "Resumo Profissional*",
-        experiencia: "Experiência Profissional",
-        formacao: "Formação Acadêmica",
-        habilidades: "Habilidades Técnicas",
-        idiomas: "Idiomas",
-        certificacoes: "Certificações/Cursos",
-        dataCertificacao: "Data",
-        cargaHoraria: "Carga Horária",
-        descricaoCertificacao: "Descrição",
-        tipoFormacao: "Tipo de Formação",
-        curso: "Curso",
-        instituicao: "Instituição",
-        periodo: "Período",
-        cargo: "Cargo",
-        empresa: "Empresa",
-        tecnologias: "Tecnologias utilizadas",
-        atividades: "Atividades realizadas",
-        resultados: "Resultados alcançados (com métricas)",
-        idioma: "Idioma",
-        nivel: "Nível",
-        certificacao: "Certificação",
-      },
-      placeholders: {
-        nome: "Ex: João da Silva",
-        cargoDesejado: "Ex: Desenvolvedor Front-end React",
-        ddd: "Ex: 11",
-        telefone: "Ex: 99999-9999",
-        cidade: "Ex: São Paulo, SP",
-        linkedin: "seuperfil",
-        portfolio: "Ex: github.com/seuuser",
-        email: "Ex: seuemail@exemplo.com",
-        resumo:
-          "Ex: Desenvolvedor Front-end com 5 anos de experiência em React e TypeScript. Especializado em criar interfaces de usuário responsivas e acessíveis...",
-        curso: "Ex: Bacharelado em Ciência da Computação",
-        instituicao: "Ex: Universidade de São Paulo",
-        periodo: "Ex: 2015 - 2019",
-        cargo: "Ex: Desenvolvedor Front-end React",
-        empresa: "Ex: Google Inc.",
-        tecnologias: "Ex: React, TypeScript, Redux, Node.js",
-        atividades: "Descreva suas responsabilidades (1 item por linha)",
-        resultados: "Ex: Reduzi o tempo de carregamento em 40% através de...",
-        habilidades:
-          "Ex: JavaScript, React, Node.js, HTML/CSS, Git, AWS, Docker",
-        idioma: "Ex: Inglês",
-        certificacao: "Ex: Certificação AWS Cloud Practitioner",
-        linkedin: "seuperfil",
-        github: "seu-usuario",
-        gitlab: "seu-usuario",
-        behance: "seu-usuario",
-        portfolio: "seusite.com",
-        outro: "URL completo",
-      },
-      botoes: {
-        adicionarExperiencia: "Adicionar Experiência",
-        adicionarFormacao: "Adicionar Formação",
-        adicionarIdioma: "Adicionar Idioma",
-        adicionarCertificacao: "Adicionar Certificação",
-        gerarCV: "Gerar Currículo em PDF",
-      },
-      mensagens: {
-        nenhumaExperiencia: "Nenhuma experiência adicionada (opcional)",
-        nenhumIdioma: "Nenhum idioma adicionado (opcional)",
-        nenhumaCertificacao: "Nenhuma certificação adicionada (opcional)",
-        sucesso: "Currículo gerado com sucesso!",
-        gerando: "Gerando Currículo...",
-      },
-      secoesPDF: {
-        resumo: "RESUMO",
-        experiencia: "EXPERIÊNCIA",
-        formacao: "FORMAÇÃO",
-        habilidades: "HABILIDADES",
-        idiomas: "IDIOMAS",
-        certificacoes: "CERTIFICAÇÕES",
-      },
-      niveisIdioma: [
-        "Básico",
-        "Intermediário",
-        "Avançado",
-        "Fluente",
-        "Nativo",
-      ],
+  pt: {
+    tituloApp: "Gerador de Currículo",
+    subtituloApp: "Crie um currículo profissional em minutos",
+    dicasATS: "Dicas para um currículo ATS-friendly",
+    dicasLista: [
+      "Use palavras-chave relevantes para a vaga",
+      "Mantenha o formato simples e legível",
+      "Inclua métricas e resultados concretos",
+      "Destaque suas principais habilidades técnicas",
+    ],
+    campos: {
+      nome: "Informações Pessoais",
+      cargoDesejado: "Cargo Desejado",
+      codigoPais: "Código do País",
+      ddd: "DDD",
+      telefone: "Telefone",
+      cidade: "Cidade",
+      linkedin: "LinkedIn",
+      portfolio: "Portfolio/GitHub",
+      email: "Email*",
+      resumo: "Resumo Profissional*",
+      experiencia: "Experiência Profissional",
+      formacao: "Formação Acadêmica",
+      habilidades: "Habilidades Técnicas",
+      idiomas: "Idiomas",
+      certificacoes: "Certificações/Cursos",
+      dataCertificacao: "Data",
+      cargaHoraria: "Carga Horária",
+      descricaoCertificacao: "Descrição",
+      tipoFormacao: "Tipo de Formação",
+      curso: "Curso",
+      instituicao: "Instituição",
+      periodo: "Período",
+      cargo: "Cargo",
+      empresa: "Empresa",
+      tecnologias: "Tecnologias utilizadas",
+      atividades: "Atividades realizadas",
+      resultados: "Resultados alcançados (com métricas)",
+      idioma: "Idioma",
+      nivel: "Nível",
+      certificacao: "Certificação",
     },
-    en: {
-      tituloApp: "Resume Generator",
-      subtituloApp: "Create a professional resume in minutes",
-      dicasATS: "Tips for an ATS-friendly resume",
-      dicasLista: [
-        "Use relevant keywords for the position",
-        "Keep the format simple and readable",
-        "Include metrics and concrete results",
-        "Highlight your main technical skills",
-      ],
-      campos: {
-        nome: "Full Name*",
-        cargoDesejado: "Desired Position",
-        codigoPais: "Country Code",
-        ddd: "Area Code",
-        telefone: "Phone",
-        cidade: "City",
-        linkedin: "LinkedIn",
-        portfolio: "Portfolio/GitHub",
-        email: "Email*",
-        resumo: "Professional Summary*",
-        experiencia: "Professional Experience",
-        formacao: "Education",
-        habilidades: "Technical Skills",
-        idiomas: "Languages",
-        certificacoes: "Certifications",
-        tipoFormacao: "Education Level",
-        curso: "Course",
-        instituicao: "Institution",
-        periodo: "Period",
-        cargo: "Position",
-        empresa: "Company",
-        tecnologias: "Technologies used",
-        atividades: "Responsibilities",
-        resultados: "Achievements (with metrics)",
-        idioma: "Language",
-        nivel: "Level",
-        certificacao: "Certification",
-      },
-      placeholders: {
-        nome: "Ex: John Smith",
-        cargoDesejado: "Ex: React Front-end Developer",
-        ddd: "Ex: 212",
-        telefone: "Ex: 555-123-4567",
-        cidade: "Ex: New York, NY",
-        linkedin: "yourprofile",
-        portfolio: "Ex: github.com/youruser",
-        email: "Ex: your.email@example.com",
-        resumo:
-          "Ex: Front-end Developer with 5 years of experience in React and TypeScript. Specialized in creating responsive and accessible user interfaces...",
-        curso: "Ex: Bachelor's in Computer Science",
-        instituicao: "Ex: University of São Paulo",
-        periodo: "Ex: 2015 - 2019",
-        cargo: "Ex: React Front-end Developer",
-        empresa: "Ex: Google Inc.",
-        tecnologias: "Ex: React, TypeScript, Redux, Node.js",
-        atividades: "Describe your responsibilities (1 item per line)",
-        resultados: "Ex: Reduced loading time by 40% through...",
-        habilidades:
-          "Ex: JavaScript, React, Node.js, HTML/CSS, Git, AWS, Docker",
-        idioma: "Ex: English",
-        certificacao: "Ex: AWS Cloud Practitioner Certification",
-        linkedin: "yourprofile",
-        github: "your-username",
-        gitlab: "your-username",
-        behance: "your-username",
-        portfolio: "yourwebsite.com",
-        outro: "Full URL",
-      },
-      botoes: {
-        adicionarExperiencia: "Add Experience",
-        adicionarFormacao: "Add Education",
-        adicionarIdioma: "Add Language",
-        adicionarCertificacao: "Add Certification",
-        gerarCV: "Generate PDF Resume",
-      },
-      mensagens: {
-        nenhumaExperiencia: "No experience added (optional)",
-        nenhumIdioma: "No languages added (optional)",
-        nenhumaCertificacao: "No certifications added (optional)",
-        sucesso: "Resume generated successfully!",
-        gerando: "Generating Resume...",
-      },
-      secoesPDF: {
-        resumo: "SUMMARY",
-        experiencia: "EXPERIENCE",
-        formacao: "EDUCATION",
-        habilidades: "SKILLS",
-        idiomas: "LANGUAGES",
-        certificacoes: "CERTIFICATIONS",
-      },
-      niveisIdioma: ["Basic", "Intermediate", "Advanced", "Fluent", "Native"],
+    placeholders: {
+      nome: "Ex: João da Silva",
+      cargoDesejado: "Ex: Desenvolvedor Front-end React",
+      ddd: "Ex: 11",
+      telefone: "Ex: 99999-9999",
+      cidade: "Ex: São Paulo, SP",
+      linkedin: "seuperfil",
+      portfolio: "Ex: github.com/seuuser",
+      email: "Ex: seuemail@exemplo.com",
+      resumo:
+        "Ex: Desenvolvedor Front-end com 5 anos de experiência em React e TypeScript. Especializado em criar interfaces de usuário responsivas e acessíveis...",
+      curso: "Ex: Bacharelado em Ciência da Computação",
+      instituicao: "Ex: Universidade de São Paulo",
+      periodo: "Ex: 2015 - 2019",
+      cargo: "Ex: Desenvolvedor Front-end React",
+      empresa: "Ex: Google Inc.",
+      tecnologias: "Ex: React, TypeScript, Redux, Node.js",
+      atividades: "Descreva suas responsabilidades (1 item por linha)",
+      resultados: "Ex: Reduzi o tempo de carregamento em 40% através de...",
+      habilidades:
+        "Ex: JavaScript, React, Node.js, HTML/CSS, Git, AWS, Docker",
+      idioma: "Ex: Inglês",
+      certificacao: "Ex: Certificação AWS Cloud Practitioner",
+      linkedin: "seuperfil",
+      github: "seu-usuario",
+      gitlab: "seu-usuario",
+      behance: "seu-usuario",
+      portfolio: "seusite.com",
+      outro: "URL completo",
     },
-    es: {
-      tituloApp: "Generador de Currículum",
-      subtituloApp: "Crea un currículum profesional en minutos",
-      dicasATS: "Consejos para un currículum compatible con ATS",
-      dicasLista: [
-        "Usa palabras clave relevantes para el puesto",
-        "Mantén el formato simple y legible",
-        "Incluye métricas y resultados concretos",
-        "Destaca tus principales habilidades técnicas",
-      ],
-      campos: {
-        nome: "Nombre Completo*",
-        cargoDesejado: "Puesto Deseado",
-        codigoPais: "Código de País",
-        ddd: "Código de Área",
-        telefone: "Teléfono",
-        cidade: "Ciudad",
-        linkedin: "LinkedIn",
-        portfolio: "Portfolio/GitHub",
-        email: "Email*",
-        resumo: "Resumen Profesional*",
-        experiencia: "Experiencia Profesional",
-        formacao: "Formación Académica",
-        habilidades: "Habilidades Técnicas",
-        idiomas: "Idiomas",
-        certificacoes: "Certificaciones",
-        tipoFormacao: "Nivel de Formación",
-        curso: "Curso",
-        instituicao: "Institución",
-        periodo: "Período",
-        cargo: "Puesto",
-        empresa: "Empresa",
-        tecnologias: "Tecnologías utilizadas",
-        atividades: "Responsabilidades",
-        resultados: "Logros (con métricas)",
-        idioma: "Idioma",
-        nivel: "Nivel",
-        certificacao: "Certificación",
-      },
-      placeholders: {
-        nome: "Ej: Juan Pérez",
-        cargoDesejado: "Ej: Desarrollador Front-end React",
-        ddd: "Ej: 11",
-        telefone: "Ej: 99999-9999",
-        cidade: "Ej: Madrid, España",
-        linkedin: "tuperfil",
-        portfolio: "Ej: github.com/tuusuario",
-        email: "Ej: tu.email@ejemplo.com",
-        resumo:
-          "Ej: Desarrollador Front-end con 5 años de experiencia en React y TypeScript. Especializado en crear interfaces de usuario responsivas y accesibles...",
-        curso: "Ej: Licenciatura en Informática",
-        instituicao: "Ej: Universidad de São Paulo",
-        periodo: "Ej: 2015 - 2019",
-        cargo: "Ej: Desarrollador Front-end React",
-        empresa: "Ej: Google Inc.",
-        tecnologias: "Ej: React, TypeScript, Redux, Node.js",
-        atividades: "Describe tus responsabilidades (1 ítem por línea)",
-        resultados: "Ej: Reduje el tiempo de carga en 40% mediante...",
-        habilidades:
-          "Ej: JavaScript, React, Node.js, HTML/CSS, Git, AWS, Docker",
-        idioma: "Ej: Inglés",
-        certificacao: "Ej: Certificación AWS Cloud Practitioner",
-        linkedin: "superfil",
-        github: "su-usuario",
-        gitlab: "su-usuario",
-        behance: "su-usuario",
-        portfolio: "sutitio.com",
-        outro: "URL completa",
-      },
-      botoes: {
-        adicionarExperiencia: "Añadir Experiencia",
-        adicionarFormacao: "Añadir Formación",
-        adicionarIdioma: "Añadir Idioma",
-        adicionarCertificacao: "Añadir Certificación",
-        gerarCV: "Generar CV en PDF",
-      },
-      mensagens: {
-        nenhumaExperiencia: "Ninguna experiencia añadida (opcional)",
-        nenhumIdioma: "Ningún idioma añadido (opcional)",
-        nenhumaCertificacao: "Ninguna certificación añadida (opcional)",
-        sucesso: "¡Currículum generado con éxito!",
-        gerando: "Generando Currículum...",
-      },
-      secoesPDF: {
-        resumo: "RESUMEN",
-        experiencia: "EXPERIENCIA",
-        formacao: "FORMACIÓN",
-        habilidades: "HABILIDADES",
-        idiomas: "IDIOMAS",
-        certificacoes: "CERTIFICACIONES",
-      },
-      niveisIdioma: ["Básico", "Intermedio", "Avanzado", "Fluido", "Nativo"],
+    botoes: {
+      adicionarExperiencia: "Adicionar Experiência",
+      adicionarFormacao: "Adicionar Formação",
+      adicionarIdioma: "Adicionar Idioma",
+      adicionarCertificacao: "Adicionar Certificação",
+      gerarCV: "Gerar Currículo em PDF",
     },
-  };
+    mensagens: {
+      nenhumaExperiencia: "Nenhuma experiência adicionada (opcional)",
+      nenhumIdioma: "Nenhum idioma adicionado (opcional)",
+      nenhumaCertificacao: "Nenhuma certificação adicionada (opcional)",
+      sucesso: "Currículo gerado com sucesso!",
+      gerando: "Gerando Currículo...",
+    },
+    secoesPDF: {
+      resumo: "RESUMO",
+      experiencia: "EXPERIÊNCIA",
+      formacao: "FORMAÇÃO",
+      habilidades: "HABILIDADES",
+      idiomas: "IDIOMAS",
+      certificacoes: "CERTIFICAÇÕES",
+    },
+    niveisIdioma: [
+      "Básico",
+      "Intermediário",
+      "Avançado",
+      "Fluente",
+      "Nativo",
+    ],
+  },
+  en: {
+    tituloApp: "Resume Generator",
+    subtituloApp: "Create a professional resume in minutes",
+    dicasATS: "Tips for an ATS-friendly resume",
+    dicasLista: [
+      "Use relevant keywords for the position",
+      "Keep the format simple and readable",
+      "Include metrics and concrete results",
+      "Highlight your main technical skills",
+    ],
+    campos: {
+      nome: "Full Name*",
+      cargoDesejado: "Desired Position",
+      codigoPais: "Country Code",
+      ddd: "Area Code",
+      telefone: "Phone",
+      cidade: "City",
+      linkedin: "LinkedIn",
+      portfolio: "Portfolio/GitHub",
+      email: "Email*",
+      resumo: "Professional Summary*",
+      experiencia: "Professional Experience",
+      formacao: "Education",
+      habilidades: "Technical Skills",
+      idiomas: "Languages",
+      certificacoes: "Certifications",
+      tipoFormacao: "Education Level",
+      curso: "Course",
+      instituicao: "Institution",
+      periodo: "Period",
+      cargo: "Position",
+      empresa: "Company",
+      tecnologias: "Technologies used",
+      atividades: "Responsibilities",
+      resultados: "Achievements (with metrics)",
+      idioma: "Language",
+      nivel: "Level",
+      certificacao: "Certification",
+    },
+    placeholders: {
+      nome: "Ex: John Smith",
+      cargoDesejado: "Ex: React Front-end Developer",
+      ddd: "Ex: 212",
+      telefone: "Ex: 555-123-4567",
+      cidade: "Ex: New York, NY",
+      linkedin: "yourprofile",
+      portfolio: "Ex: github.com/youruser",
+      email: "Ex: your.email@example.com",
+      resumo:
+        "Ex: Front-end Developer with 5 years of experience in React and TypeScript. Specialized in creating responsive and accessible user interfaces...",
+      curso: "Ex: Bachelor's in Computer Science",
+      instituicao: "Ex: University of São Paulo",
+      periodo: "Ex: 2015 - 2019",
+      cargo: "Ex: React Front-end Developer",
+      empresa: "Ex: Google Inc.",
+      tecnologias: "Ex: React, TypeScript, Redux, Node.js",
+      atividades: "Describe your responsibilities (1 item per line)",
+      resultados: "Ex: Reduced loading time by 40% through...",
+      habilidades:
+        "Ex: JavaScript, React, Node.js, HTML/CSS, Git, AWS, Docker",
+      idioma: "Ex: English",
+      certificacao: "Ex: AWS Cloud Practitioner Certification",
+      linkedin: "yourprofile",
+      github: "your-username",
+      gitlab: "your-username",
+      behance: "your-username",
+      portfolio: "yourwebsite.com",
+      outro: "Full URL",
+    },
+    botoes: {
+      adicionarExperiencia: "Add Experience",
+      adicionarFormacao: "Add Education",
+      adicionarIdioma: "Add Language",
+      adicionarCertificacao: "Add Certification",
+      gerarCV: "Generate PDF Resume",
+    },
+    mensagens: {
+      nenhumaExperiencia: "No experience added (optional)",
+      nenhumIdioma: "No languages added (optional)",
+      nenhumaCertificacao: "No certifications added (optional)",
+      sucesso: "Resume generated successfully!",
+      gerando: "Generating Resume...",
+    },
+    secoesPDF: {
+      resumo: "SUMMARY",
+      experiencia: "EXPERIENCE",
+      formacao: "EDUCATION",
+      habilidades: "SKILLS",
+      idiomas: "LANGUAGES",
+      certificacoes: "CERTIFICATIONS",
+    },
+    niveisIdioma: ["Basic", "Intermediate", "Advanced", "Fluent", "Native"],
+  },
+  es: {
+    tituloApp: "Generador de Currículum",
+    subtituloApp: "Crea un currículum profesional en minutos",
+    dicasATS: "Consejos para un currículum compatible con ATS",
+    dicasLista: [
+      "Usa palabras clave relevantes para el puesto",
+      "Mantén el formato simple y legible",
+      "Incluye métricas y resultados concretos",
+      "Destaca tus principales habilidades técnicas",
+    ],
+    campos: {
+      nome: "Nombre Completo*",
+      cargoDesejado: "Puesto Deseado",
+      codigoPais: "Código de País",
+      ddd: "Código de Área",
+      telefone: "Teléfono",
+      cidade: "Ciudad",
+      linkedin: "LinkedIn",
+      portfolio: "Portfolio/GitHub",
+      email: "Email*",
+      resumo: "Resumen Profesional*",
+      experiencia: "Experiencia Profesional",
+      formacao: "Formación Académica",
+      habilidades: "Habilidades Técnicas",
+      idiomas: "Idiomas",
+      certificacoes: "Certificaciones",
+      tipoFormacao: "Nivel de Formación",
+      curso: "Curso",
+      instituicao: "Institución",
+      periodo: "Período",
+      cargo: "Puesto",
+      empresa: "Empresa",
+      tecnologias: "Tecnologías utilizadas",
+      atividades: "Responsabilidades",
+      resultados: "Logros (con métricas)",
+      idioma: "Idioma",
+      nivel: "Nivel",
+      certificacao: "Certificación",
+    },
+    placeholders: {
+      nome: "Ej: Juan Pérez",
+      cargoDesejado: "Ej: Desarrollador Front-end React",
+      ddd: "Ej: 11",
+      telefone: "Ej: 99999-9999",
+      cidade: "Ej: Madrid, España",
+      linkedin: "tuperfil",
+      portfolio: "Ej: github.com/tuusuario",
+      email: "Ej: tu.email@ejemplo.com",
+      resumo:
+        "Ej: Desarrollador Front-end con 5 años de experiencia en React y TypeScript. Especializado en crear interfaces de usuario responsivas y accesibles...",
+      curso: "Ej: Licenciatura en Informática",
+      instituicao: "Ej: Universidad de São Paulo",
+      periodo: "Ej: 2015 - 2019",
+      cargo: "Ej: Desarrollador Front-end React",
+      empresa: "Ej: Google Inc.",
+      tecnologias: "Ej: React, TypeScript, Redux, Node.js",
+      atividades: "Describe tus responsabilidades (1 ítem por línea)",
+      resultados: "Ej: Reduje el tiempo de carga en 40% mediante...",
+      habilidades:
+        "Ej: JavaScript, React, Node.js, HTML/CSS, Git, AWS, Docker",
+      idioma: "Ej: Inglés",
+      certificacao: "Ej: Certificación AWS Cloud Practitioner",
+      linkedin: "superfil",
+      github: "su-usuario",
+      gitlab: "su-usuario",
+      behance: "su-usuario",
+      portfolio: "sutitio.com",
+      outro: "URL completa",
+    },
+    botoes: {
+      adicionarExperiencia: "Añadir Experiencia",
+      adicionarFormacao: "Añadir Formación",
+      adicionarIdioma: "Añadir Idioma",
+      adicionarCertificacao: "Añadir Certificación",
+      gerarCV: "Generar CV en PDF",
+    },
+    mensagens: {
+      nenhumaExperiencia: "Ninguna experiencia añadida (opcional)",
+      nenhumIdioma: "Ningún idioma añadido (opcional)",
+      nenhumaCertificacao: "Ninguna certificación añadida (opcional)",
+      sucesso: "¡Currículum generado con éxito!",
+      gerando: "Generando Currículum...",
+    },
+    secoesPDF: {
+      resumo: "RESUMEN",
+      experiencia: "EXPERIENCIA",
+      formacao: "FORMACIÓN",
+      habilidades: "HABILIDADES",
+      idiomas: "IDIOMAS",
+      certificacoes: "CERTIFICACIONES",
+    },
+    niveisIdioma: ["Básico", "Intermedio", "Avanzado", "Fluido", "Nativo"],
+  },
+};
 
 export const paisesTelefone = [
-    { codigo: "+55", nome: "Brasil (+55)" },
-    { codigo: "+1", nome: "EUA/Canadá (+1)" },
-    { codigo: "+54", nome: "Argentina (+54)" },
-    { codigo: "+351", nome: "Portugal (+351)" },
-    { codigo: "+34", nome: "Espanha (+34)" },
-    { codigo: "+49", nome: "Alemanha (+49)" },
-    { codigo: "+33", nome: "França (+33)" },
-    { codigo: "+44", nome: "Reino Unido (+44)" },
-    { codigo: "+39", nome: "Itália (+39)" },
-    { codigo: "+61", nome: "Austrália (+61)" },
-  ];
+  { codigo: "+55", nome: "Brasil (+55)" },
+  { codigo: "+1", nome: "EUA/Canadá (+1)" },
+  { codigo: "+54", nome: "Argentina (+54)" },
+  { codigo: "+351", nome: "Portugal (+351)" },
+  { codigo: "+34", nome: "Espanha (+34)" },
+  { codigo: "+49", nome: "Alemanha (+49)" },
+  { codigo: "+33", nome: "França (+33)" },
+  { codigo: "+44", nome: "Reino Unido (+44)" },
+  { codigo: "+39", nome: "Itália (+39)" },
+  { codigo: "+61", nome: "Austrália (+61)" },
+];
 
 export const tiposCurso = [
-    {
-      valor: "superior",
-      label: "Ensino Superior em",
-      label_en: "Bachelor's Degree in",
-      label_es: "Licenciatura en",
-    },
-    {
-      valor: "tecnologo",
-      label: "Tecnólogo em",
-      label_en: "Technology Degree in",
-      label_es: "Tecnólogo en",
-    },
-    {
-      valor: "medio",
-      label: "Ensino Médio",
-      label_en: "High School",
-      label_es: "Bachillerato",
-    },
-    {
-      valor: "tecnico",
-      label: "Curso Técnico em",
-      label_en: "Technical Course in",
-      label_es: "Curso Técnico en",
-    },
-    {
-      valor: "pos",
-      label: "Pós-Graduação em",
-      label_en: "Postgraduate in",
-      label_es: "Posgrado en",
-    },
-    {
-      valor: "mestrado",
-      label: "Mestrado em",
-      label_en: "Master's Degree in",
-      label_es: "Máster en",
-    },
-    {
-      valor: "doutorado",
-      label: "Doutorado em",
-      label_en: "PhD in",
-      label_es: "Doctorado en",
-    },
+  {
+    valor: "superior",
+    label: "Ensino Superior em",
+    label_en: "Bachelor's Degree in",
+    label_es: "Licenciatura en",
+  },
+  {
+    valor: "tecnologo",
+    label: "Tecnólogo em",
+    label_en: "Technology Degree in",
+    label_es: "Tecnólogo en",
+  },
+  {
+    valor: "medio",
+    label: "Ensino Médio",
+    label_en: "High School",
+    label_es: "Bachillerato",
+  },
+  {
+    valor: "tecnico",
+    label: "Curso Técnico em",
+    label_en: "Technical Course in",
+    label_es: "Curso Técnico en",
+  },
+  {
+    valor: "pos",
+    label: "Pós-Graduação em",
+    label_en: "Postgraduate in",
+    label_es: "Posgrado en",
+  },
+  {
+    valor: "mestrado",
+    label: "Mestrado em",
+    label_en: "Master's Degree in",
+    label_es: "Máster en",
+  },
+  {
+    valor: "doutorado",
+    label: "Doutorado em",
+    label_en: "PhD in",
+    label_es: "Doctorado en",
+  },
+];
+
+export const sanitizeForATS = (text) => {
+  if (!text) return text;
+  if (typeof text !== "string") return "";
+
+  // Mapeamento de caracteres especiais para suas versões ASCII
+  const replacements = {
+    "∑": "SUM",
+    "≤": "<=",
+    "≥": ">=",
+    "≠": "!=",
+    "―": "-",
+    "–": "-",
+    "—": "-",
+    "“": '"',
+    "”": '"',
+    "‘": "'",
+    "’": "'",
+    "…": "...",
+    "•": "-",
+    "→": "->",
+    "←": "<-",
+    "±": "+/-",
+    μ: "u",
+    "°": "deg",
+  };
+
+  // Substitui caracteres especiais e remove outros não suportados
+  return text.replace(
+    /[∑≤≥≠―–—“”‘’…•→←±μ°]/g,
+    (char) => replacements[char] || ""
+  );
+};
+
+  export const meses = [
+    { valor: "01", label: "Janeiro" },
+    { valor: "02", label: "Fevereiro" },
+    { valor: "03", label: "Março" },
+    { valor: "04", label: "Abril" },
+    { valor: "05", label: "Maio" },
+    { valor: "06", label: "Junho" },
+    { valor: "07", label: "Julho" },
+    { valor: "08", label: "Agosto" },
+    { valor: "09", label: "Setembro" },
+    { valor: "10", label: "Outubro" },
+    { valor: "11", label: "Novembro" },
+    { valor: "12", label: "Dezembro" },
   ];
 
-  export const sanitizeForATS = (text) => {
-    if (!text) return text;
-    if (typeof text !== "string") return "";
-  
-    // Mapeamento de caracteres especiais para suas versões ASCII
-    const replacements = {
-      "∑": "SUM",
-      "≤": "<=",
-      "≥": ">=",
-      "≠": "!=",
-      "―": "-",
-      "–": "-",
-      "—": "-",
-      "“": '"',
-      "”": '"',
-      "‘": "'",
-      "’": "'",
-      "…": "...",
-      "•": "-",
-      "→": "->",
-      "←": "<-",
-      "±": "+/-",
-      μ: "u",
-      "°": "deg",
-    };
-  
-    // Substitui caracteres especiais e remove outros não suportados
-    return text.replace(
-      /[∑≤≥≠―–—“”‘’…•→←±μ°]/g,
-      (char) => replacements[char] || ""
-    );
-  };
+  export const statusFormacao = [
+    { valor: "completo", label: "Completo" },
+    { valor: "andamento", label: "Em andamento" },
+    { valor: "trancado", label: "Trancado" },
+  ];
+
+  export const tiposRedesSociais = [
+    { valor: "linkedin", label: "LinkedIn", prefixo: "linkedin.com/in/" },
+    { valor: "github", label: "GitHub", prefixo: "github.com/" },
+    { valor: "gitlab", label: "GitLab", prefixo: "gitlab.com/" },
+    { valor: "behance", label: "Behance", prefixo: "behance.net/" },
+    { valor: "portfolio", label: "Portfolio", prefixo: "" },
+    { valor: "outro", label: "outro", prefixo: "" },
+  ];
